[REF] Put the purge_records_more_than_reserve into a context class.
