+++ bin/migrate.py
@@ -244,7 +244,7 @@ def cmd_system(command):
 def disable_dangerous_stuff():
     queries = [
         ("UPDATE ir_mail_server SET active = FALSE, smtp_host = 'f'", True),
-        ("UPDATE fetchmail_server SETE active = FALSE", False),
+        ("UPDATE fetchmail_server SET active = FALSE", False),
         ("UPDATE ir_cron SET active = FALSE", True),
     ]
     dbname = os.environ["PGDATABASE"]
